requirements.txt
asgiref==3.11.1
Django==5.2.12
pillow==12.1.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
python-dotenv==1.0.0
markdown==3.5.1
gunicorn==21.2.0
